requirements.txt
mysql-connector-python
matplotlib
numpy
Pillow
tkinter-tooltip
captcha
scikit-learn
pandas
joblib
